extract shared editor form for create + edit event
